requirements.txt
sqlalchemy==1.4.39
fastapi==0.101.1
pydantic==2.1.1
uvicorn==0.23.2
python-jose[cryptography]
passlib[bcrypt]